Create AutomateEmail
A simple yet powerful technique to send emails without having to set up or configure a mail client is to just leverage the mail client on the host of a user.
Let's say we were able to pivot on a machine and found an interesting file that we want to exfiltrate. Here is a neat PowerShell script that leverages COM automation. As I described
earlier, this can be achieved with any scripting language that can leverage COM. You can put the following commands into a file and run them with powershell.exe, or run the commands individually on the PowerShell command line: